Alpha was separated
