+++ Data/qtdb/qrs_params.txt
@@ -1,10 +1,10 @@
 WDC_SCALE_ID 3
 WDC_SCALE_ID_AUX 4
-CANDIDATE_WINDOW_MIN_BEAT 0.280
-CANDIDATE_WINDOW_TRAINING_PERIOD 8.0
-CANDIDATE_BEATS_COUNT_TRAINING_PERIOD 4
-CANDIDATE_WINDOW_DELTA 0.100
-CANDIDATE_THRESHOLD_DELTA 0.5
+ALPHA_MIN_BEAT 0.280
+ALPHA_TRAINING_WINDOW 8.0
+ALPHA_BEATS_IN_TRAINING_WINDOW 4
+ALPHA_QRS_WINDOW 0.100
+ALPHA_THRESHOLD 0.5
 ONSET_WINDOW 0.100
 ONSET_THRESHOLD_MM 0.055
 ONSET_AMPLITUDE_DECREASING 0.3
