+++ Source/Model/main/delineation/qrs/delineation.py
@@ -9,13 +9,9 @@
     qrs_zcs - список пересечений нуля детализирующими вейвлет-коэффициентами.
 """
 
-from Source.Model.main.delineation.qrs.onset import *
-from Source.Model.main.delineation.qrs.offset import *
-from Source.Model.main.delineation.qrs.peak import *
-from Source.Model.main.delineation.qrs.zcs import *
-from Source.Model.main.delineation.qrs.routines import *
-from Source.Model.main.params.qrs import *
 from Source.Model.main.delineation.qrs.gamma.morphology import *
+from Source.Model.main.delineation.qrs.onset import *
+from Source.Model.main.delineation.qrs.peak import *
 
 
 class InvalidQRSDelineation(Exception):
