+++ Source/Model/main/params/qrs.py
@@ -5,11 +5,11 @@ QRSParams = dict()
 QRSParams['WDC_SCALE_ID'] = QRSDefaultParams.WDC_SCALE_ID
 QRSParams['WDC_SCALE_ID_AUX'] = QRSDefaultParams.WDC_SCALE_ID_AUX
 
-QRSParams['CANDIDATE_WINDOW_MIN_BEAT'] = QRSDefaultParams.CANDIDATE_WINDOW_MIN_BEAT
-QRSParams['CANDIDATE_WINDOW_DELTA'] = QRSDefaultParams.CANDIDATE_WINDOW_DELTA
-QRSParams['CANDIDATE_WINDOW_TRAINING_PERIOD'] = QRSDefaultParams.CANDIDATE_WINDOW_TRAINING_PERIOD
-QRSParams['CANDIDATE_BEATS_COUNT_TRAINING_PERIOD'] = QRSDefaultParams.CANDIDATE_BEATS_COUNT_TRAINING_PERIOD
-QRSParams['CANDIDATE_THRESHOLD_DELTA'] = QRSDefaultParams.CANDIDATE_THRESHOLD_DELTA
+QRSParams['ALPHA_MIN_BEAT'] = QRSDefaultParams.ALPHA_MIN_BEAT
+QRSParams['ALPHA_TRAINING_WINDOW'] = QRSDefaultParams.ALPHA_TRAINING_WINDOW
+QRSParams['ALPHA_BEATS_IN_TRAINING_WINDOW'] = QRSDefaultParams.ALPHA_BEATS_IN_TRAINING_WINDOW
+QRSParams['ALPHA_QRS_WINDOW'] = QRSDefaultParams.ALPHA_QRS_WINDOW
+QRSParams['ALPHA_THRESHOLD'] = QRSDefaultParams.ALPHA_THRESHOLD
 
 QRSParams['ONSET_WINDOW'] = QRSDefaultParams.ONSET_WINDOW
 QRSParams['ONSET_THRESHOLD_MM'] = QRSDefaultParams.ONSET_THRESHOLD_MM
